requirements.txt
pandas>=2.0.0
anthropic>=0.25.0
python-dotenv>=1.0.0
fuzzywuzzy>=0.18.0
python-levenshtein>=0.21.0
dataclasses-json>=0.5.0
typing-extensions>=4.0.0 